moved enums and structs into own files
