+++ Komodo/Assets/Packages/KomodoCore/Runtime/Scripts/RuntimeSession/EntityState.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Komodo.Runtime
+{
+    [System.Serializable]
+    public struct EntityState
+    {
+        public int id;
+        public Position latest;
+        public bool render;
+        public bool locked;
+    }
+}
