+++ Komodo/Assets/Packages/KomodoCore/Runtime/Scripts/RuntimeSession/EntityState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c28255cb2cf4a9d489bbee5f880787ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
