+++ Komodo/Assets/Packages/KomodoCore/Runtime/Scripts/RuntimeSession/EntityType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 553819b7d55679f4c9783217f1282e16
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
