+++ Komodo/Assets/Packages/KomodoCore/Runtime/Scripts/RuntimeSession/InteractionType.cs
@@ -1,0 +1,29 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Komodo.Runtime
+{
+    public enum INTERACTIONS
+    {
+        LOOK = 0,
+        LOOK_END = 1,
+        SHOW = 2,
+        HIDE = 3,
+        GRAB = 4,
+        DROP = 5,
+        CHANGE_SCENE = 6,
+        SLICE_OBJECT = 7,
+        LOCK = 8,
+        UNLOCK = 9,
+        LINE = 10,
+        LINE_END = 11,
+        SHOW_MENU = 12,
+        HIDE_MENU = 13,
+
+        SETTING_TAB = 14,
+        PEOPLE_TAB = 15,
+        INTERACTION_TAB = 16,
+        CREATE_TAB = 17,
+    }
+}
