+++ Komodo/Assets/Packages/KomodoCore/Runtime/Scripts/RuntimeSession/InteractionType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5ecb7ba8a86c8d94cb4c7064cc5ae17f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
