+++ Komodo/Assets/Packages/KomodoCore/Runtime/Scripts/RuntimeSession/SessionState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ccbf976550777ce4b9dfdb9dfe8290f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
