+++ Komodo/Assets/Packages/KomodoCore/Runtime/Scripts/RuntimeSession/SpeechToText.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a2f972960dd2da247b6706c5ee5773ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
